+++ src/hooks/useProduct.ts
@@ -1,8 +1,11 @@
-import { useState } from "react";
-import { useSelector } from "react-redux";
+import { useState, useEffect } from "react";
+import { useDispatch, useSelector } from "react-redux";
 import { RootState } from "../modules/Store";
+import { getProducts } from "../modules/Products";
 
 function useProduct() {
+  const dispatch = useDispatch();
+
   const [limit, setLimit] = useState(8);
   const [page, setPage] = useState(1);
 
@@ -14,6 +17,10 @@ function useProduct() {
     window.innerWidth >= 1440 && setLimit(8);
     window.innerWidth < 1440 && setLimit(6);
   };
+
+  useEffect(() => {
+    dispatch(getProducts());
+  }, [dispatch]);
 
   return { limit, page, setPage, products, handleLimit };
 }
