+++ src/modules/Carts.ts
@@ -1,6 +1,5 @@
 import { createSlice, createAsyncThunk } from "@reduxjs/toolkit";
 import axios from "axios";
-
 import { Cart, Product } from "../types/dto";
 
 interface CartsState {
@@ -17,6 +16,17 @@ const cartsSlice = createSlice({
   reducers: {},
   extraReducers: (builder) => {
     builder
+      .addCase(getCarts.pending, (state) => {
+        state.cartsList = [...state.cartsList];
+      })
+      .addCase(getCarts.fulfilled, (state, action) => {
+        state.cartsList = [...action.payload];
+      })
+      .addCase(getCarts.rejected, (_, action) => {
+        console.log(action.error.message);
+      });
+
+    builder
       .addCase(postCart.pending, (state) => {
         state.cartsList = [...state.cartsList];
       })
@@ -27,6 +37,15 @@ const cartsSlice = createSlice({
         console.log(action.error.message);
       });
   },
+});
+
+export const getCarts = createAsyncThunk("cart/get", async () => {
+  const response = await axios({
+    method: "get",
+    url: "http://localhost:3003/carts",
+    headers: { "Cache-Control": "no-cache" },
+  });
+  return response.data;
 });
 
 export const postCart = createAsyncThunk(
